[#74] Prevent to submit data when already have saved data in DB
