+++ backend/tests/test_071_unlimited_member_survey.py
@@ -45,7 +45,7 @@ class TestCreateUnlimitedMemberQuestionnaire():
                              client: AsyncClient) -> None:
         res = await client.post(
             app.url_path_for("data:create", form_id=1, submitted=0),
-            params={"locked_by": 1},
+            params={"locked_by": 2},
             json=[{
                 "question": 1,
                 "repeat_index": 0,
@@ -56,11 +56,11 @@ class TestCreateUnlimitedMemberQuestionnaire():
         assert res.status_code == 200
         res = res.json()
         assert res == {
-            "id": 4,
+            "id": 2,
             "form": 1,
             "name": "",
             "geo": None,
-            "locked_by": 1,
+            "locked_by": 2,
             "created": today,
             "created_by": "Galih",
             "organisation": org_name,
@@ -90,8 +90,8 @@ class TestCreateUnlimitedMemberQuestionnaire():
             "created": today,
             "created_by": "Galih",
             "form": 1,
-            "id": 4,
-            "locked_by": 1,
+            "id": 2,
+            "locked_by": 2,
             "name": f"{org_name} - Galih - {today}",
             "organisation": org_name,
         }
@@ -111,6 +111,87 @@ class TestCreateUnlimitedMemberQuestionnaire():
             ),
             headers={"Authorization": f"Bearer {account.token}"})
         assert res.status_code == 200
+
+    @pytest.mark.asyncio
+    async def test_submit_data(self, app: FastAPI, session: Session,
+                               client: AsyncClient) -> None:
+        res = await client.post(
+            app.url_path_for("data:create", form_id=1, submitted=1),
+            params={"locked_by": 2},
+            json=[{
+                "question": 1,
+                "repeat_index": 0,
+                "comment": None,
+                "value": "Option 1"
+            }, {
+                "question": 2,
+                "repeat_index": 0,
+                "comment": None,
+                "value": "Direct submit"
+            }, {
+                "question": 3,
+                "repeat_index": 0,
+                "comment": None,
+                "value": "Female"
+            }, {
+                "question": 4,
+                "repeat_index": 0,
+                "comment": None,
+                "value": 35
+            }, {
+                "question": 5,
+                "repeat_index": 0,
+                "comment": "Q5 comment",
+                "value": 55
+            }],
+            headers={"Authorization": f"Bearer {account.token}"})
+        assert res.status_code == 200
+        res = res.json()
+        assert res == {
+            "id": 3,
+            "form": 1,
+            "name": "Direct submit",
+            "geo": None,
+            "locked_by": 2,
+            "created": today,
+            "created_by": "Galih",
+            "organisation": org_name,
+            "submitted_by": "Galih",
+            "updated": today,
+            "submitted": today,
+            "answer": [
+                {
+                    "comment": None,
+                    "question": 1,
+                    "repeat_index": 0,
+                    "value": "Option 1"
+                },
+                {
+                    "comment": None,
+                    "question": 2,
+                    "repeat_index": 0,
+                    "value": "Direct submit"
+                },
+                {
+                    "comment": None,
+                    "question": 3,
+                    "repeat_index": 0,
+                    "value": "Female"
+                },
+                {
+                    "comment": None,
+                    "question": 4,
+                    "repeat_index": 0,
+                    "value": 35
+                },
+                {
+                    "comment": "Q5 comment",
+                    "question": 5,
+                    "repeat_index": 0,
+                    "value": 55
+                }
+            ]
+        }
 
     @pytest.mark.asyncio
     async def test_get_third_webform_from_bucket(
